feat: Implement DAG boilerplate
The result of a brainstorming session to come up with a minimal and
performant DAG implementation that is only concerned with CX (CNOT)
gates.

The goal of the DAG implementation is to create a DAG than can easily be
moved between languages.

As a first step, the DAG should condense the qiskit-DAG down to a
representation that includes only information relevant to addressing the
qubit mapping problem.